requirements.txt
# requirements.txt

# Core Libraries
playwright

# Configuration
python-dotenv
configparser

# Testing
pytest
pytest-mock